Extract field deconstruction compiler fixture
